code folder renamed to BORO standards
