+++ ifc_source/ClassRegistry.py
