+++ ifc_source/GenericDefinition.py
@@ -1,4 +1,4 @@
-from Ifc.ClassRegistry import ifc_fallback_definition
+from ifc_source.ClassRegistry import ifc_fallback_definition
 
 
 @ifc_fallback_definition
