+++ ifc_source/Type.py
@@ -2,7 +2,7 @@
 import os
 # END EDIT
 
-from Ifc.ClassRegistry import ifc_definition
+from ifc_source.ClassRegistry import ifc_definition
 
 
 @ifc_definition
